restructured packages and misc refactorings
