+++ pkg/etcdaccess/option.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package client
+package etcdaccess
 
 // Options contains options used by the client.
 type Options struct {
